Moved no-unpickling import into fake_torch.py and updated convert_model.py to use it as well as accept an arguement to run the restricted unpickler instead, which has been merged into this branch.
